requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
python-decouple
django-filter
